# api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  description: "The Monika Cloud Cloud Platform API provides web services for the Monika suite of apps and systems."
  version: "0.1.0"
  title: "Monika Cloud Patform API"
  termsOfService: "https://cloud.monika.com/terms/"
  contact:
    email: "swood@monika.com"
  license:
    name: "Commercial"
    url: "http://cloud.monika.com/licenses/license-1.0.html"
host: "127.0.0.1:9000"
basePath: "/api/v1"
...
schemes:
- "http"
paths:
  /user:
    x-swagger-router-controller: user.controller
    post:
      tags:
      - "user"
      summary: "Add a new user to the system"
      description: "Add a new user to the system"
      operationId: "addUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: body
        name: body
        description: "The user object to be added"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "User successfuly added"
          schema:
            type: object
            properties:
              userId:
                type: string
                format: uuid
        400:
          description: "The user data posted was not valid"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "There is a user already matching that data"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
  /user/list:
    x-swagger-router-controller: user.controller
    post:
      tags:
      - "user"
      summary: "Get a list of users, filtered and sorted"
      description: "Get a list of users, filtered and sorted"
      operationId: "listUsers"
      parameters:
      - in: body
        name: body
        description: "pararmeters for filtering and sorting the list of users"
        required: false
        schema:
          type: object
          properties:
            userListFilterQuery:
              $ref: "#/definitions/UserListFilterQuery"
            userListSortQuery:
              $ref: "#/definitions/UserListSortQuery"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/User"
        400:
          description: "Invalid query data"
          schema:
            $ref: "#/definitions/Error"
  /user/{userId}:
    x-swagger-router-controller: user.controller
    get:
      tags:
      - "user"
      summary: "Retreive the details of the user by id"
      description: "Retreive the details of the user by id"
      operationId: "getUser"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: path
        description: "The id of the user account you would like to activate"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        404:
          description: "User not found"
          schema:
            $ref: "#/definitions/Error"
    patch:
      tags:
      - "user"
      summary: "Update an existing user's details in the system"
      description: "Updates the users details"
      operationId: "updateUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: path
        description: "The id of the user you would like to update"
        required: true
        type: string
        format: uuid
      - name: JsonPatch
        in: body
        required: true
        schema:
          $ref: "#/definitions/PatchRequest"
      responses:
        200:
          description: "User successfuly updated"
          schema:
            type: string
        400:
          description: "The user data posted was not valid"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "User not found"
          schema:
            $ref: "#/definitions/Error"
  /user/{userId}/activate:
    x-swagger-router-controller: user.controller
    patch:
      tags:
      - "user"
      summary: "Activates a user that was previously deactivated"
      description: "Activates a user that was previously deactivated"
      operationId: "activateUser"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: path
        description: "The id of the user account you would like to activate"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "successful operation"
        400:
          description: "User already active"
        404:
          description: "User not found"
  /user/{userId}/deactivate:
    x-swagger-router-controller: user.controller
    patch:
      tags:
      - "user"
      summary: "Deactivates a user that was previously deactivated"
      description: "A user cannot be deleted from the system, because there may be audit records assocaited with them. Instead, we deactivate the user."
      operationId: "deactivateUser"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: path
        description: "The id of the user account you would like to deactivate"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "successful operation"
        400:
          description: "User already deactivated"
        404:
          description: "User not found"
  /user/login:
    x-swagger-router-controller: user.controller
    post:
      tags:
      - "user"
      summary: "Logs user into the system"
      description: ""
      operationId: "loginUser"
      produces:
      - "application/json"
      parameters:
      - name: body
        in: body
        description: "The login information"
        required: true
        schema:
          type: object
          properties:
            email:
              type: string
              description: "The email address of the user logging in"
            password:
              type: string
              description: "The password for login in clear text"
          required:
          - email
          - password
      responses:
        200:
          description: "successful operation"
          schema:
            type: string
          headers:
            X-Rate-Limit:
              type: integer
              format: "int32"
              description: "calls per hour allowed by the user"
            X-Expires-After:
              type: string
              format: "date-time"
              description: "date in UTC when token expires"
        400:
          description: "Invalid username/password supplied"
...
  /user/{userId}/2fa:
    x-swagger-router-controller: user.controller
    delete:
      tags:
      - "user"
      summary: "Deletes or resets the users 2fa key"
      description: "To reset a users 2FA, we delete their 2FA key so that upon next login, it asks the user to setup 2FA again"
      operationId: "userDelete2fa"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: path
        description: "The id of the user who's 2FA service you would like to delete or reset"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "User's 2FA Key was successfully deleted."
        404:
          description: "User not found"
    get:
      tags:
      - "user"
      summary: "Get the current 2FA settings (and key) for the user or create one if it doesn't exist"
      description: "If the user does not have 2FA configured, when they log in they will be prompted to set it up. This call returns the key which can be used on a key generation app such as Authy, Google Authenticator, etc"
      operationId: "userNew2faKey"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: path
        description: "The id of the user who's 2FA service you would generate and retrieve a new key for"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "User's 2FA Key was successfully deleted."
          schema:
            $ref: "#/definitions/User2faSettings"
        404:
          description: "User not found"
  /user/{userId}/2fa/activate:
    x-swagger-router-controller: user.controller
    patch:
      tags:
      - "user"
      summary: "Activate the current 2FA key for the user"
      description: "To activate the 2FA key for the user, a valid 2FA token must be received."
      operationId: "userActivate2fa"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: path
        description: "The id of the user who's 2FA service you would like to activate"
        required: true
        type: string
        format: uuid
      - name: "2FAToken"
        in: body
        description: "The generated Token as at the current date/time"
        required: true
        schema:
          type: string
          format: password
      responses:
        200:
          description: "Success. The 2FA service for the user was activated"
        404:
          description: "User not found"
  /user/{userId}/2fa/verify:
    x-swagger-router-controller: user.controller
    post:
      tags:
      - "user"
      summary: "Send the 2FA token and verify the session"
      description: "After validating email and password, the 2FA token must be provided to validate the users session. This call receives the 2FA token and validates the users session."
      operationId: "userVerify2fa"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: path
        description: "The id of the user who's 2FA service you would like to activate"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "Success. The 2FA service for the user was activated"
        404:
          description: "User not found"
  /user/password/reset/email:
    x-swagger-router-controller: user.controller
    post:
      tags:
      - "user"
      summary: "Request a password reset code be sent to your email address"
      description: "If the email address is found, a password reset code will be sent."
      operationId: "userPasswordResetSendCodeToEmail"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "emailAddress"
        in: body
        description: "the email address of the user"
        required: true
        schema: 
          type: string
          format: "email"
      responses:
        200:
          description: "Success. A password reset token has been emailed to the users email address."
        404:
          description: "A user with that email address was not found."
  /user/password/reset/check:
    x-swagger-router-controller: user.controller
    get:
      tags:
      - "user"
      summary: "check that the password reset code is valid"
      description: "If the email address is found, a password reset code will be sent."
      operationId: "userPasswordResetCheckCode"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "passwordResetCode"
        in: body
        description: "the password reset code"
        required: true
        schema:
          type: string
          format: password
      responses:
        200:
          description: "Success. The password reset code is valid."
        404:
          description: "A password reset request with that code was not found."
  /user/password/reset/change:
    x-swagger-router-controller: user.controller
    post:
      tags:
      - "user"
      summary: "As part of the password reset process, change the users password using the reset code"
      description: "As part of the password reset process, change the users password using the reset code"
      operationId: "userPasswordResetChangePassword"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "passwordResetInformation"
        in: body
        description: "The information required for resetting the password"
        required: true
        schema:
          type: object
          required:
          - passwordResetCode
          - password
          - confirmPassword
          properties: 
            passwordResetCode:
              type: string
              format: password
            password:
              type: string
              format: password
            confirmPassword:
              type: string
              format: password
      responses:
        200:
          description: "Success. The user's password has been changed"
        400:
          description: "Failed. The passwords do not match. Please try again."
        404:
          description: "A password reset request with that code was not found."
...
definitions:
...
  Error:
    type: object
    properties:
      code:
        type: integer
      message:
        type: string
    required:
    - code
    - message
...
  PatchRequest:
    type: array
    items:
      $ref: "#/definitions/PatchDocument"
  PatchDocument: 
    description: A JSONPatch document as defined by RFC 6902 
    required:
     - "op"
     - "path"
    properties: 
     op: 
      type: string 
      description: The operation to be performed 
      enum:
       - "add"
       - "remove"
       - "replace"
       - "move"
       - "copy"
       - "test"
     path: 
      type: string 
      description: A JSON-Pointer 
     value: 
      type: string 
      description: The value to be used within the operations.
     from: 
      type: string 
      description: A string containing a JSON Pointer value.
  User:
    type: object
    required:
    - "email"
    - "ownership"
    properties:
      id:
        type: string
        format: uuid
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
        format: "email"
      password:
        type: string
      phone:
        type: string
      status:
        type: string
        description: "Status"
        enum:
        - "active"
        - "deactivated"
      companyName:
        type: string
      profilePic:
        type: string
        format : file
      roleId:
        type: string
        format: uuid
      roleName:
        type: string
      ownership:
        type: object
        properties:
          ownershipType:
            type: string
            description: "What type of object the user account is owned by"
            enum:
            - "monika"
            - "none"
            - "tenant"
            - "region"
            - "site"
          tenantId:
            type: string
            format: uuid
          tenantName:
            type: string
          regionId:
            type: string
            format: uuid
          regionName:
            type: string
          siteId:
            type: string
            format: uuid
          siteName:
            type: string
      created:
        type: string
        format: date-time
      lastModified:
        type: string
        format: date-time
  UserListSortQuery:
    type: object
    properties:
      sortBy:
        type: string
        description: "Sort By"
        enum:
        - "tenant name"
        - "region name"
        - "site name"
        - "role name"
        - "company name"
        - "date created"
        - "first name"
        - "last name"
        - "full name"
        - "email address"
      order:
        type: string
        description: "Sort Order"
        enum:
        - "ascending"
        - "descending"
  UserListFilterQuery:
    type: object
    properties:
      tenantId:
        type: string
        format: uuid
      regionId:
        type: string
        format: uuid
      siteId:
        type: string
        format: uuid
      roleId:
        type: string
        format: uuid
      status:
        type: string
        description: "Status"
        enum:
        - "active"
        - "deactivated"
      keyword:
        type: string
        description: "perform a keyword search and only show those results with matches"
  User2faSettings:
    type: object
    properties:
      id:
        type: string
        format: uuid
      userId:
        type: string
        format: uuid
      key:
        type: string
        format: password
      status:
        type: string
        description: "2FA Status"
        enum:
        - "inactive"
        - "active"
...